A little more factoring around layout
